renaming files and modifying bazel scripts
